01_discovery/requirements.txt
# Discovery dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
dnspython>=2.0.0
ddgs>=9.0.0
requests>=2.31.0
PyYAML>=6.0.0
